Re-organize package structure
