+++ src/main/java/com/github/dakusui/pcond/core/multi/MultiFunction.java
@@ -1,17 +1,24 @@
 package com.github.dakusui.pcond.core.multi;
 
-import com.github.dakusui.pcond.core.currying.FormattingUtils;
-import com.github.dakusui.pcond.core.identifieable.PrintableMultiFunction;
+import com.github.dakusui.pcond.core.currying.CurryingUtils;
+import com.github.dakusui.pcond.core.printable.PrintableFunction;
 
 import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Objects;
 import java.util.function.Function;
+import java.util.function.Supplier;
 
 import static com.github.dakusui.pcond.core.currying.Checks.validateParamOrderList;
-import static com.github.dakusui.pcond.core.currying.FormattingUtils.formatMethodName;
-import static com.github.dakusui.pcond.core.currying.ReflectionsUtils.invokeStaticMethod;
+import static com.github.dakusui.pcond.core.refl.ReflUtils.formatMethodName;
+import static com.github.dakusui.pcond.core.refl.ReflUtils.invokeStaticMethod;
+import static com.github.dakusui.pcond.internals.InternalChecks.requireArgumentListSize;
 import static com.github.dakusui.pcond.internals.InternalChecks.requireStaticMethod;
 import static java.util.Arrays.asList;
+import static java.util.Objects.requireNonNull;
+import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toList;
 
 /**
@@ -21,17 +28,6 @@ import static java.util.stream.Collectors.toList;
  * @param <R>
  */
 public interface MultiFunction<R> extends Function<List<? super Object>, R> {
-  static <R> PrintableMultiFunction<R> createFromStaticMethod(Method method, List<Integer> paramOrder) {
-    validateParamOrderList(paramOrder, method.getParameterCount());
-    requireStaticMethod(method);
-    return new PrintableMultiFunction.Builder<R>(args -> invokeStaticMethod(method, (paramOrder).stream().map(args::get).toArray()))
-        .name(method.getName())
-        .formatter(() -> formatMethodName(method) + FormattingUtils.formatParameterOrder(paramOrder))
-        .addParameters(paramOrder.stream().map(i -> method.getParameterTypes()[i]).collect(toList()))
-        .identityArgs(asList(method, validateParamOrderList(paramOrder, method.getParameterCount())))
-        .$();
-  }
-
   /**
    * Returns a name of this function.
    *
@@ -60,4 +56,96 @@ public interface MultiFunction<R> extends Function<List<? super Object>, R> {
    * @return The type of the returned value.
    */
   Class<? extends R> returnType();
+
+  class Impl<R> extends PrintableFunction<List<? super Object>, R> implements MultiFunction<R> {
+    private final String         name;
+    private final List<Class<?>> parameterTypes;
+
+    protected Impl(
+        Object creator,
+        List<Object> args,
+        Supplier<String> formatter,
+        String name,
+        Function<? super List<? super Object>, ? extends R> function,
+        List<Class<?>> parameterTypes) {
+      super(
+          creator,
+          args,
+          formatter,
+          function);
+      this.name = name;
+      this.parameterTypes = new ArrayList<>(parameterTypes);
+    }
+
+    @Override
+    public String name() {
+      return name;
+    }
+
+    @Override
+    public int arity() {
+      return parameterTypes.size();
+    }
+
+    @Override
+    public Class<?> parameterType(int i) {
+      return parameterTypes.get(i);
+    }
+
+  }
+
+  /**
+   * A builder for a {@link MultiFunction} instance.
+   *
+   * @param <R> The type of value returned by the multi-function built by this object.
+   */
+  class Builder<R> {
+    private final Object                                              creator        = Builder.class;
+    private       List<Object>                                        identityArgs;
+    private       String                                              name           = "(anonymous)";
+    private final Function<? super List<? super Object>, ? extends R> body;
+    private final List<Class<?>>                                      parameterTypes = new LinkedList<>();
+    private       Supplier<String>                                    formatter      = () -> name + "(" + parameterTypes.stream().map(Class::getSimpleName).collect(joining(",")) + ")";
+
+    public Builder(Function<List<Object>, R> body) {
+      requireNonNull(body);
+      this.body = args -> body.apply(requireArgumentListSize(requireNonNull(args), parameterTypes.size()));
+    }
+
+    public Builder<R> addParameters(List<Class<?>> parameterTypes) {
+      requireNonNull(parameterTypes).stream().map(this::addParameter).forEach(Objects::requireNonNull);
+      return this;
+    }
+
+    public Builder<R> identityArgs(List<Object> identity) {
+      this.identityArgs = requireNonNull(identity);
+      return this;
+    }
+
+    public Builder<R> name(String name) {
+      this.name = name;
+      return this;
+    }
+
+    public Builder<R> addParameter(Class<?> parameterType) {
+      this.parameterTypes.add(requireNonNull(parameterType));
+      return this;
+    }
+
+    public Builder<R> formatter(Supplier<String> formatter) {
+      this.formatter = requireNonNull(formatter);
+      return this;
+    }
+
+    public MultiFunction<R> $() {
+      return new Impl<R>(
+          this.creator,
+          this.identityArgs,
+          this.formatter,
+          this.name,
+          this.body,
+          this.parameterTypes
+      );
+    }
+  }
 }
