+++ src/main/java/com/github/dakusui/pcond/functions/Printables.java
@@ -1,8 +1,8 @@
 package com.github.dakusui.pcond.functions;
 
 
-import com.github.dakusui.pcond.core.identifieable.IdentifiableFunctionFactory;
-import com.github.dakusui.pcond.core.identifieable.IdentifiablePredicateFactory;
+import com.github.dakusui.pcond.core.printable.PrintableFunctionFactory;
+import com.github.dakusui.pcond.core.printable.PrintablePredicateFactory;
 
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -14,29 +14,19 @@ import java.util.function.Supplier;
 public enum Printables {
   ;
 
-  public static <T> Predicate<T> predicate(String s, Predicate<T> predicate) {
-    return Printables.predicate(() -> s, predicate);
+  public static <T> Predicate<T> predicate(String name, Predicate<T> predicate) {
+    return PrintablePredicateFactory.leaf(name, predicate);
   }
 
-  public static <T> Predicate<T> predicate(Supplier<String> s, Predicate<T> predicate) {
-    return Printables.printablePredicate(s.get(), predicate);
+  public static <T> Predicate<T> predicate(Supplier<String> formatter, Predicate<T> predicate) {
+    return PrintablePredicateFactory.leaf(formatter, predicate);
   }
 
-  public static <T> Predicate<T> printablePredicate(String s, Predicate<T> predicate) {
-    return IdentifiablePredicateFactory.leaf(Printables.class, () -> s, predicate);
+  public static <T, R> Function<T, R> function(String name, Function<T, R> function) {
+    return PrintableFunctionFactory.function(name, function);
   }
 
-  public static <T, R> Function<T, R> function(String s, Function<T, R> function) {
-    return Printables.printableFunction(s, function);
-  }
-
-
-  public static <T, R> Function<T, R> function(Supplier<String> s, Function<T, R> function) {
-    return Printables.printableFunction(s.get(), function);
-  }
-
-  @SuppressWarnings("unchecked")
-  public static <T, R> Function<T, R> printableFunction(String s, Function<? super T, ? extends R> function) {
-    return (Function<T, R>) IdentifiableFunctionFactory.function(() -> s, function);
+  public static <T, R> Function<T, R> function(Supplier<String> formatter, Function<T, R> function) {
+    return PrintableFunctionFactory.function(formatter, function);
   }
 }
