+++ src/test/java/com/github/dakusui/pcond/ut/TransformingPredicateTest.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.pcond.ut;
 
-import com.github.dakusui.pcond.core.identifieable.IdentifiableFunctionFactory;
-import com.github.dakusui.pcond.core.identifieable.IdentifiablePredicateFactory;
+import com.github.dakusui.pcond.core.printable.PrintableFunctionFactory;
+import com.github.dakusui.pcond.core.printable.PrintablePredicateFactory;
 import com.github.dakusui.pcond.functions.Functions;
 import com.github.dakusui.pcond.functions.Predicates;
 import com.github.dakusui.pcond.utils.ut.TestBase;
@@ -56,8 +56,8 @@ public class TransformingPredicateTest extends TestBase {
     Predicate<String> equalToHello = Predicates.isEqualTo("hello");
     Predicate<String> p = Predicates.transform(identity).check(equalToHello);
 
-    Predicate<String> q = Predicates.transform(IdentifiableFunctionFactory.function(() -> "", identity))
-        .check(IdentifiablePredicateFactory.leaf(() -> "", equalToHello));
+    Predicate<String> q = Predicates.transform(PrintableFunctionFactory.function(() -> "", identity))
+        .check(PrintablePredicateFactory.leaf(() -> "", equalToHello));
 
     assertEquals(
         p.hashCode(),
@@ -66,14 +66,14 @@ public class TransformingPredicateTest extends TestBase {
 
   @Test
   public void givenNullForName$whenToString$thenLooksGood() {
-    IdentifiablePredicateFactory.TransformingPredicate<Object, Object> p = new IdentifiablePredicateFactory.TransformingPredicate<>(null, Predicates.alwaysTrue(), Functions.identity());
+    PrintablePredicateFactory.TransformingPredicate<Object, Object> p = new PrintablePredicateFactory.TransformingPredicate<>(null, Predicates.alwaysTrue(), Functions.identity());
     System.out.println(p.toString());
     assertEquals("identity alwaysTrue", p.toString());
   }
 
   @Test
   public void givenNonNullName$whenToString$thenLooksGood() {
-    IdentifiablePredicateFactory.TransformingPredicate<Object, Object> p = new IdentifiablePredicateFactory.TransformingPredicate<>("hello->", Predicates.alwaysTrue(), Functions.identity());
+    PrintablePredicateFactory.TransformingPredicate<Object, Object> p = new PrintablePredicateFactory.TransformingPredicate<>("hello->", Predicates.alwaysTrue(), Functions.identity());
     System.out.println(p.toString());
     assertEquals("hello->identity alwaysTrue", p.toString());
   }
